nftminter/backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
google-genai
python-dotenv
Pillow
ipfshttpclient
